chore: refactor frontend file widget to use the new backend file
