+++ bciers/apps/administration/app/components/operations/OperationInformationPage.tsx
@@ -19,6 +19,8 @@ const OperationInformationPage = async ({
     operation = await getOperationWithDocuments(operationId);
   } else throw new Error(`Invalid operation id: ${operationId}`);
 
+  console.log(operation);
+
   if (operation?.error) throw new Error("Error fetching operation information");
 
   const formSchema = await createAdministrationOperationInformationSchema(
