+++ bciers/libs/components/src/form/theme/readOnlyTheme.ts
@@ -15,7 +15,6 @@ import {
   BoroIdWidget,
   OptedOutOperationWidget,
   OperationRepresentativeWidget,
-  NewFileWidget,
 } from "../widgets";
 import ApplyComplianceUnitsWidget from "@/compliance/src/app/widgets/ApplyComplianceUnitsWidget";
 import {
@@ -34,7 +33,6 @@ const readOnlyTheme = {
     DateWidget: ReadOnlyDateWidget,
     EmailWidget: ReadOnlyWidget,
     FileWidget: ReadOnlyFileWidget,
-    NewFileWidget: NewFileWidget,
     MultiSelectWidgetWithTooltip: ReadOnlyMultiSelectWidget,
     RegulatedProductMultiSelectWidget: ReadOnlyMultiSelectWidget,
     PhoneWidget: ReadOnlyWidget,
