+++ bciers/libs/components/src/form/widgets/FileWidget.tsx
@@ -1,78 +1,11 @@
-"use client";
-
-import { useCallback, useEffect, useRef, useState } from "react";
-import type { ChangeEvent } from "react";
-import {
-  dataURItoBlob,
-  FormContextType,
-  Registry,
-  RJSFSchema,
-  StrictRJSFSchema,
-  TranslatableString,
-  WidgetProps,
-} from "@rjsf/utils";
+import { actionHandler } from "@bciers/actions";
 import { AlertIcon } from "@bciers/components/icons";
 import { useSessionRole } from "@bciers/utils/src/sessionUtils";
+import { TranslatableString, WidgetProps } from "@rjsf/utils";
+import { useRef } from "react";
 
-const addNameToDataURL = (dataURL: string, name: string) => {
-  if (dataURL === null) {
-    return null;
-  }
-  return dataURL.replace(";base64", `;name=${encodeURIComponent(name)};base64`);
-};
-
+const MAX_FILE_SIZE = 20000000;
 type FileScanStatus = "Unscanned" | "Clean" | "Quarantined";
-
-const getScanStatusFromDataURL = (dataURL: string | null): FileScanStatus => {
-  if (dataURL === null) {
-    return "Unscanned";
-  }
-  const scanStatus = dataURL.match(/scanstatus=([^;]+)/)?.[1];
-  if (!scanStatus) {
-    return "Unscanned";
-  }
-  return scanStatus as FileScanStatus;
-};
-
-type FileInfoType = {
-  dataURL?: string | null;
-  name: string;
-  size: number;
-  type: string;
-  scanStatus?: FileScanStatus | null;
-};
-
-const processFile = (file: File): Promise<FileInfoType> => {
-  const { name, size, type } = file;
-  return new Promise((resolve, reject) => {
-    const reader = new window.FileReader();
-    reader.onerror = reject;
-    reader.onload = (event) => {
-      if (typeof event.target?.result === "string") {
-        resolve({
-          dataURL: addNameToDataURL(event.target.result, name),
-          name,
-          size,
-          type,
-          scanStatus: getScanStatusFromDataURL(event.target.result),
-        });
-      } else {
-        resolve({
-          dataURL: null,
-          name,
-          size,
-          type,
-          scanStatus: null,
-        });
-      }
-    };
-    reader.readAsDataURL(file);
-  });
-};
-
-const processFiles = (files: FileList) => {
-  return Promise.all(Array.from(files).map(processFile));
-};
 
 // Show a different message depending on the fileScanStatus
 const showScanStatus = (status: FileScanStatus | undefined | null) => {
@@ -101,153 +34,107 @@ const showScanStatus = (status: FileScanStatus | undefined | null) => {
   return null;
 };
 
-function FileInfoPreview<
-  T = any,
-  S extends StrictRJSFSchema = RJSFSchema,
-  F extends FormContextType = any,
->({
-  fileInfo,
-  registry,
-}: {
-  readonly fileInfo: FileInfoType;
-  readonly registry: Registry<T, S, F>;
-}) {
-  const { translateString } = registry;
-  const { dataURL, name } = fileInfo;
-  if (!dataURL) {
-    return null;
-  }
+const useFileUploadWidget = (
+  endpoint: string,
+  method: "PUT" | "PATCH" | "POST",
+  pathToRevalidate?: string,
+): [typeof formContext, typeof submitWithFiles] => {
+  const files: Record<string, File> = {};
 
-  return (
-    <>
-      {" "}
-      <a download={`preview-${name}`} href={dataURL} className="file-download">
-        {translateString(TranslatableString.PreviewLabel)}
-      </a>
-    </>
-  );
-}
+  const formContext = {
+    onFileSelected: (file: File, propName: string) => {
+      files[propName] = file;
+    },
+  };
 
-export function FilesInfo<
-  T = any,
-  S extends StrictRJSFSchema = RJSFSchema,
-  F extends FormContextType = any,
->({
-  filesInfo,
+  // Proxy for the actionHandler, using a multipart/form-data encoding instead
+  const submitWithFiles = async (formData: any) => {
+    const formDataWithFiles = new FormData();
+
+    formDataWithFiles.append("payload", JSON.stringify(formData));
+
+    for (const [propName, file] of Object.entries(files)) {
+      formDataWithFiles.append(propName, file);
+    }
+
+    return await actionHandler(endpoint, method, pathToRevalidate, {
+      body: formDataWithFiles,
+    });
+  };
+
+  return [formContext, submitWithFiles];
+};
+
+export function FileElement({
+  value,
   preview,
-  registry,
 }: {
-  readonly filesInfo: FileInfoType[];
+  readonly value: {
+    id: number;
+    name: string;
+    status?: FileScanStatus;
+  };
   readonly preview?: boolean;
-  readonly registry: Registry<T, S, F>;
 }) {
-  if (filesInfo.length === 0) {
-    return null;
-  }
+  const handlePreview = () => {
+    // Similarly to the report attachments, we'll call the backend
+    // to fetch a temporary URL for the file in GCS, and open it in a new tab
+  };
+
   return (
     <ul className="m-0 py-0 flex flex-col justify-start list-none">
-      {filesInfo.map((fileInfo) => {
-        const { name, scanStatus } = fileInfo;
-        const isQuarantined = scanStatus === "Quarantined";
-        return (
-          <li key={name} data-name={name}>
-            {showScanStatus(scanStatus) || name}
-            {preview && !isQuarantined && (
-              <FileInfoPreview<T, S, F>
-                fileInfo={fileInfo}
-                registry={registry}
-              />
-            )}
-          </li>
-        );
-      })}
+      <li data-name={value.name}>
+        {JSON.stringify(value)} {showScanStatus(value.status) || value.name}{" "}
+        {preview && value.status !== "Quarantined" && (
+          <button className="button-link file-download" onClick={handlePreview}>
+            {TranslatableString.PreviewLabel}
+          </button>
+        )}
+      </li>
     </ul>
   );
 }
 
-export const extractFileInfo = (dataURLs: string[]): FileInfoType[] => {
-  return dataURLs
-    .filter((dataURL) => dataURL)
-    .map((dataURL) => {
-      const { blob, name } = dataURItoBlob(dataURL);
-      const scanStatus = getScanStatusFromDataURL(dataURL);
-      return {
-        dataURL,
-        name: name,
-        size: blob.size,
-        type: blob.type,
-        scanStatus,
-      };
-    });
-};
+const FileWidget: React.FC<WidgetProps> = (props) => {
+  const {
+    id,
+    required,
+    disabled,
+    readonly,
+    options,
+    onChange,
+    name,
+    registry,
+    value,
+  } = props;
 
-const FileWidget = ({
-  id,
-  disabled,
-  readonly,
-  required,
-  multiple,
-  onChange,
-  value,
-  options,
-  registry,
-}: WidgetProps) => {
-  // We need to store the value in state to prevent loosing the value when user switches between tabs
-  const [localValue, setLocalValue] = useState(value);
-  const [filesInfo, setFilesInfo] = useState<FileInfoType[]>(
-    Array.isArray(localValue)
-      ? extractFileInfo(localValue)
-      : extractFileInfo([localValue]),
-  );
-  // 🥷 Prevent resetting the value to null when user switch tabs
-  useEffect(() => {
-    if (localValue && !value) {
-      onChange(localValue);
-    }
-  }, [localValue, onChange, value]);
+  if (!registry.formContext?.onFileSelected) {
+    throw new Error(
+      "FileWidget is being used without the useFileUploadWidget context. File uploads will not work.",
+    );
+  }
 
   const role = useSessionRole();
   const isCasInternal = role?.includes("cas") && !role?.includes("pending");
 
-  const hiddenFileInput = useRef<HTMLInputElement>(null);
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files?.[0];
+    if (file) {
+      if (file.size > MAX_FILE_SIZE) {
+        alert("File size must be less than 20MB");
+        return;
+      }
+      registry.formContext.onFileSelected(file, name);
+    }
 
+    // From the RJSF perspective this will just be a string with the file name.
+    onChange(file?.name);
+  };
+
+  const hiddenFileInput = useRef<HTMLInputElement>(null);
   const handleClick = () => {
     hiddenFileInput.current?.click();
   };
-
-  const handleChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      event.preventDefault();
-      const maxSize = 20000000;
-      const files = event.target.files;
-      if (!files) {
-        return;
-      }
-
-      processFiles(files).then((filesInfoEvent) => {
-        const newValue = filesInfoEvent.map((fileInfo) => {
-          if (fileInfo.size > maxSize) {
-            alert("File size must be less than 20MB");
-            return;
-          }
-          return fileInfo.dataURL;
-        });
-        if (multiple) {
-          setFilesInfo(filesInfo.concat(filesInfoEvent[0]));
-          onChange(localValue.concat(newValue[0]));
-          setLocalValue(localValue.concat(newValue[0]));
-        } else {
-          setFilesInfo(filesInfoEvent);
-          onChange(newValue[0]);
-          setLocalValue(newValue[0]);
-        }
-      });
-    },
-    [multiple, localValue, filesInfo, onChange],
-  );
-
-  const disabledColour =
-    disabled || readonly ? "text-bc-bg-dark-grey" : "text-bc-link-blue";
 
   const uploadLabel = options.uploadLabel
     ? String(options.uploadLabel)
@@ -256,7 +143,9 @@ const FileWidget = ({
     ? String(options.reuploadLabel)
     : "Reupload attachment";
 
-  /*   File input styling options are limited so we are attaching a ref to it, hiding it and triggering it with a styled button. */
+  const disabledColour =
+    disabled || readonly ? "text-bc-bg-dark-grey" : "text-bc-link-blue";
+
   return (
     <div className="py-4 flex flex-col md:flex-row items-start gap-2 md:gap-4">
       {!isCasInternal && (
@@ -266,7 +155,7 @@ const FileWidget = ({
           className={`p-0 decoration-solid border-0 text-base md:text-lg bg-transparent cursor-pointer underline flex-shrink-0 ${disabledColour}`}
           style={{ whiteSpace: "nowrap" }}
         >
-          {localValue ? reuploadLabel : uploadLabel}
+          {value ? reuploadLabel : uploadLabel}
         </button>
       )}
       <input
@@ -283,12 +172,8 @@ const FileWidget = ({
         value=""
         accept={options.accept ? String(options.accept) : undefined}
       />
-      {localValue ? (
-        <FilesInfo
-          registry={registry}
-          filesInfo={filesInfo}
-          preview={options.filePreview}
-        />
+      {value ? (
+        <FileElement value={value} preview={options.filePreview} />
       ) : (
         <span className="text-base md:text-lg flex-shrink-0">
           No attachment was uploaded.
@@ -298,4 +183,4 @@ const FileWidget = ({
   );
 };
 
-export default FileWidget;
+export { FileWidget as default, useFileUploadWidget };
