+++ bciers/libs/components/src/form/widgets/readOnly/ReadOnlyFileWidget.tsx
@@ -1,38 +1,13 @@
 "use client";
 
 import { WidgetProps } from "@rjsf/utils";
-import { useEffect, useState } from "react";
-import {
-  extractFileInfo,
-  FilesInfo,
-} from "@bciers/components/form/widgets/FileWidget";
+import { FileElement } from "../FileWidget";
 
-const ReadOnlyFileWidget: React.FC<WidgetProps> = ({
-  id,
-  options,
-  registry,
-  value,
-}) => {
-  const [fileInfo, setFileInfo] = useState<any[]>([]);
-  const fileValue = Array.isArray(value) ? value : [value];
-
-  useEffect(() => {
-    // Fix for window not being defined on page load
-    if (fileValue.length) {
-      // eslint-disable-next-line react-hooks/set-state-in-effect
-      setFileInfo(fileValue);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
+const ReadOnlyFileWidget: React.FC<WidgetProps> = ({ id, options, value }) => {
   return (
     <div id={id} className="read-only-widget">
       {value ? (
-        <FilesInfo
-          filesInfo={extractFileInfo(fileInfo)}
-          preview={options.filePreview}
-          registry={registry}
-        />
+        <FileElement value={value} preview={options?.filePreview} />
       ) : (
         <>No attachment was uploaded.</>
       )}
